backend/requirements.txt
Flask
flask-cors
pymongo
dnspython
requests
python-dotenv
